Move Handlers to a classlib
