+++ TerminalGame.RelayServer/TerminalGame.RelayServer.Lib/MessageHandlerWithoutYield.cs
@@ -1,17 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Json;
 using TerminalGame.RelayServer.Domain;
 
-namespace TerminalGame.RelayServer.Benchmark
+namespace TerminalGame.RelayServer.Lib
 {
-    public class MessageHandler
+    public class MessageHandlerWithoutYield
     {
+        private long TotalRead { get; set; } = 0;
+        private long CurrentRead { get; set; } = 0;
         private bool MessageIncomplete { get; set; }
+        private int MessageSize { get; set; } = 0;
+
         private string AccumulateMessage { get; set; } = "";
 
-
+        
         public IEnumerable<SocketMessage?> DecodeBuffer(byte[] buffer, long offset, long size)
         {
             AccumulateMessage += Encoding.UTF8.GetString(buffer, (int) offset, (int) size);
